Import tags from Trello
We don't bring over colors at the moment because it's not supported by
the API, and the tags might already have colors set in Aha!. If the
tag colors are edited in Aha! after import, future imports will use
the correct tag colors.

If a Trello tag has no name, it is given a name named after the color.